Rename django/flask modules to avoid clashes
The filenames of django.py and flask.py can easily accidentally clash
with the actual django and flask packages, so they really need to  be
named something else.

Here I'm just renaming "django" to "django_helpers" and "flask" to
"flask_helpers".
