+++ tests/fixtures/django/deleted_callback_urls.py
@@ -3,13 +3,13 @@ import os
 
 # Packages
 from django.http import HttpResponseGone
-from canonicalwebteam.yaml_responses import django
+from canonicalwebteam.yaml_responses.django_helpers import create_deleted_views
 
 
 parent_dir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 
 # Add redirects
-urlpatterns = django.create_deleted_views(
+urlpatterns = create_deleted_views(
     path=f"{parent_dir}/deleted.yaml",
     view_callback=lambda request, context, settings: HttpResponseGone(
         "custom callback"
