+++ tests/fixtures/django/redirects_urls.py
@@ -4,13 +4,15 @@ import os
 # Packages
 from django.conf.urls import url
 from django.http import HttpResponse
-from canonicalwebteam.yaml_responses import django
+from canonicalwebteam.yaml_responses.django_helpers import (
+    create_redirect_views,
+)
 
 
 parent_dir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 
 # Add redirects
-urlpatterns = django.create_redirect_views(path=f"{parent_dir}/redirects.yaml")
+urlpatterns = create_redirect_views(path=f"{parent_dir}/redirects.yaml")
 
 # Standard patterns
 urlpatterns += [url("homepage", lambda request: HttpResponse("hello world"))]
